Refatorar barra de busca de atendentes
